+++ webapp/src/main/webapp/src/app/assessments/filters/exam-filters/exam-filter.service.spec.ts
@@ -23,6 +23,25 @@ describe('ExamFilterService', () => {
     fixture = new ExamFilterService();
   });
 
+  it('should filter exams by Economic Disadvantage', () => {
+    filterBy.economicDisadvantage = 1;
+
+    assessmentExam.exams[0].economicDisadvantage = true;
+    assessmentExam.exams[1].economicDisadvantage = false;
+    assessmentExam.exams[2].economicDisadvantage = false;
+    assessmentExam.exams[3].economicDisadvantage = false;
+
+    const actual = fixture.filterExams(
+      assessmentExam.exams,
+      assessmentExam.assessment,
+      filterBy
+    );
+
+    expect(actual.length).toBe(1);
+    expect(actual.some(x => x.economicDisadvantage === true)).toBeTruthy();
+    expect(actual.some(x => x.economicDisadvantage === false)).toBeFalsy();
+  });
+
   it('should filter exams by Administrative condition for IABs', () => {
     filterBy.administration = AdministrativeCondition.Standard;
     assessmentExam.assessment.type = 'iab';