+++ webapp/src/main/webapp/src/app/shared/embargo/reporting-embargo-alert.component.ts
@@ -1,13 +1,14 @@
 import { Component } from '@angular/core';
 import { ReportingEmbargoService } from './reporting-embargo.service';
 import { EmbargoAlert } from './embargo-alert.component';
+import { UserService } from '../security/service/user.service';
 
 @Component({
   selector: 'reporting-embargo-alert',
   templateUrl: 'embargo-alert.component.html'
 })
 export class ReportingEmbargoAlert extends EmbargoAlert {
-  constructor(service: ReportingEmbargoService) {
-    super(service);
+  constructor(service: ReportingEmbargoService, userService: UserService) {
+    super(service, userService);
   }
 }